Updated the circuit example model and the notebooks
