+++ tests/graph/test_graph_solve.py
@@ -20,7 +20,8 @@ def test_basic_kerbal_solve(kerbal_lpg, kerbal_random_stage_5_complete, kerbal_s
     *_, qualified_name_to_id = kerbal_stable_names
 
     rt_10_isp_id = qualified_name_to_id[
-        f"""{PARTS_LIBRARY}RT-10 "Hammer" Solid Fuel Booster::Specific Impulse: Real <<AttributeUsage>>"""
+        f"""{PARTS_LIBRARY}RT-10 "Hammer" Solid Fuel Booster"""
+        "::Specific Impulse: Real <<AttributeUsage>>"
     ]
     ft200_full_mass = qualified_name_to_id[
         f"{PARTS_LIBRARY}FL-T200 Fuel Tank::Full Mass: Real <<AttributeUsage>>"
@@ -48,7 +49,7 @@ def test_basic_kerbal_solve(kerbal_lpg, kerbal_random_stage_5_complete, kerbal_s
 
     try:
         cg1.solve_graph(kerbal_lpg)
-    except:
+    except:  # pylint: disable=bare-except
         pass
 
     for move in literal_output_moves:
@@ -74,9 +75,13 @@ def test_path_step_expression_kerbal_solve(
 
     id_to_qualified_name, qualified_name_to_id = kerbal_stable_names
 
+    assert id_to_qualified_name, "Could not find ID to qualified name mapping"
+
     pse_engine_mass = qualified_name_to_id[
         f"""{PARTS_LIBRARY}RT-10 "Hammer" Solid Fuel Booster::Specific Impulse: Real <<AttributeUsage>>"""
     ]
+
+    assert pse_engine_mass, "No PSE mass!"
 
     # incrementally step through the calculations and check progress
