+++ tests/interpretation/circuit_example/__init__.py
@@ -1,0 +1,4 @@
+from .openmdao import CircuitComponent
+from .ui import CircuitUI
+
+__all__ = ("CircuitUI", "CircuitComponent")
