+++ tests/interpretation/test_calculation_ordering.py
@@ -1,7 +1,6 @@
 import pytest
 
 from pymbe.interpretation.calc_dependencies import generate_execution_order
-from tests.conftest import kerbal_lpg, kerbal_random_stage_5_complete, kerbal_stable_names
 
 ROCKET_BUILDING = "Model::Kerbal::Rocket Building::"
 PARTS_LIBRARY = "Model::Kerbal::Parts Library::"
